requirements.txt
bert_extractive_summarizer==0.10.1
moviepy==1.0.3
openai==1.40.3
python-dotenv==1.0.1
Requests==2.32.3
summarizer==0.0.7
